api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["SuperMati"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.6"
psycopg2-binary = "^2.9.9"
strawberry-graphql = "^0.211.1"
django-extensions = "^3.2.3"
django-cors-headers = "^4.3.0"
asyncio = "^3.4.3"
pyppeteer = "^1.0.2"
pyjwt = "^2.8.0"
pydantic = "^2.4.2"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.1.0"
black = "^23.10.1"
strawberry-graphql = {extras = ["debug-server"], version = "^0.211.1"}
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
